+++ git-updater-additions.php
@@ -13,7 +13,7 @@
  * Plugin Name:       Git Updater Additions
  * Plugin URI:        https://github.com/afragen/git-updater-additions
  * Description:       Add installed repositories lacking required headers to the Git Updater plugin.
- * Version:           5.0.0
+ * Version:           5.0.0.1
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true
@@ -37,6 +37,7 @@ if ( ! defined( 'WPINC' ) ) {
 // Load Autoloader.
 require_once __DIR__ . '/vendor/autoload.php';
 
+( new GUA_Freemius() )->init();
 add_action(
 	'plugins_loaded',
 	function() {